feat: add react navigation setup
