+++ src/features/home/home-screen.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import {StyleSheet, Text, View} from 'react-native';
+
+export const HomeScreen = () => {
+  return (
+    <View style={styles.container}>
+      <Text>Home</Text>
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#234',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
